Format currency.

+++ AspMvc5Multilingual/Models/Widget.cs
@@ -4,7 +4,7 @@ namespace AspMvc5Multilingual.Models
 {
     public class Widget
     {
-        public double Money { get; set; }
+        public decimal Money { get; set; }
         public DateTime DateTime { get; set; } = DateTime.Now;
     }
 }
